Make list_index INTEGER(255)

+++ aquadoggo/migrations/20220509090252_create-operations.sql
@@ -16,7 +16,7 @@ CREATE TABLE IF NOT EXISTS operation_fields_v1 (
     name                    TEXT             NOT NULL,
     field_type              TEXT             NOT NULL,
     value                   TEXT             NULL,
-    list_index              NUMERIC          NOT NULL,
+    list_index              INTEGER(255)     NOT NULL,
     FOREIGN KEY(operation_id) REFERENCES operations_v1(operation_id)
 );
